🌱📗Docs : [강의 : 토비의 스프링부트] 섹션 VI 33_합성 애노테이션의 적용
